Start deduplicating the getSetting function
This code cannot run.
