+++ chromium/background.js
@@ -14,6 +14,8 @@
    limitations under the License.
 */
 
+import { getSetting } from "./common";
+
 if (typeof browser === 'undefined') {
     var browser = chrome;
 }
@@ -556,23 +558,4 @@ async function brieflyShowX() {
 
 async function sleep(ms) {
     return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-/**
- * getSetting gets a setting from the browser's sync storage.
- * @param {string} name - the name of the setting.
- * @param {any} default_ - the default value of the setting.
- * @returns {any}
- */
-async function getSetting(name, default_) {
-    try {
-        const v = (await browser.storage.sync.get(name))[name];
-        if (v === undefined) {
-            return default_;
-        }
-        return v;
-    } catch (err) {
-        console.error(err);
-        return default_;
-    }
 }
