+++ chromium/common.js
@@ -1,0 +1,38 @@
+/*
+   Copyright 2024 Chris Wheeler
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+if (typeof browser === 'undefined') {
+    var browser = chrome;
+}
+
+/**
+ * getSetting gets a setting from the browser's sync storage.
+ * @param {string} name - the name of the setting.
+ * @param {any} default_ - the default value of the setting.
+ * @returns {Promise<any>}
+ */
+export async function getSetting(name, default_) {
+    try {
+        const v = (await browser.storage.sync.get(name))[name];
+        if (v === undefined) {
+            return default_;
+        }
+        return v;
+    } catch (err) {
+        console.error(err);
+        return default_;
+    }
+}
