+++ chromium/options.html
@@ -100,6 +100,7 @@ as the options page. -->
       <input id="reset" type="reset" />
     </form>
 
+    <script src="common.js" type="module"></script>
     <script src="options.js"></script>
   </body>
