+++ firefox/options.html
@@ -95,6 +95,7 @@
       <input id="reset" type="reset" />
     </form>
 
+    <script src="common.js" type="module"></script>
     <script src="options.js"></script>
   </body>
